Convert resource to new style custom resource and add local system protections.
