+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'cookbooks@chef.io'
 license          'Apache 2.0'
 description      'Installs/Configures the Microsoft Web Platform Installer (WebPI)'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '2.0.0'
+version          '2.0.1'
 supports         'windows'
 depends          'windows', '>= 1.39.0'
